+++ src/main/kotlin/com/oneinch/object/Token.kt
@@ -1,5 +1,5 @@
 package com.oneinch.`object`
 
-import java.math.BigInteger
+import java.math.BigDecimal
 
-class Token(val symbol: String, val address: String, val decimals: BigInteger)
+class Token(val symbol: String, val address: String, val decimals: BigDecimal)
